chore: Refactor code and update dependencies
This commit refactors the code by adding the `BasicConfigs` class to handle basic application configurations and the `MailingConfigs` class to handle mailing configurations. It also updates the `python-dotenv` dependency to version 1.0.1 and adds the `pydantic` module with version 2.8.2. These changes were made to improve the organization of the code and ensure compatibility with the latest versions of the packages.