feat: Add user_name, first_name, last_name mappings to User model
This commit adds the `user_name`, `first_name`, and `last_name` mappings to the `User` model in the `user.py` file. These mappings allow for storing and retrieving user names and personal information in the database. By including these mappings, we enhance the functionality and data management of the user module.